requirements.txt
langchain
openai
redis
pytesseract
pillow